sync/requirements.txt
web3==7.4.0
cryptography==41.0.7
python-dotenv==1.0.1